requirements.txt
psutil
anthropic
python-dotenv
requests
kademlia